Open-Clawd/clawd-vault-master/python/pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawvault-py"
version = "0.1.0"
description = "Python SDK for ClawVault markdown memory vaults."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
  { name = "Versatly" }
]
keywords = ["ai-agent", "memory", "markdown", "obsidian", "clawvault"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "PyYAML>=6.0.2",
  "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
semantic = [
  "sentence-transformers>=3.0.0",
]
dev = [
  "pytest>=8.0.0",
]

[project.urls]
Homepage = "https://clawvault.dev"
Repository = "https://github.com/Versatly/clawvault"
Issues = "https://github.com/Versatly/clawvault/issues"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
clawvault = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
